[FEAT/#170] 화면 간 navigation 연결

+++ feature/store/src/main/java/com/napzak/market/store/store/navigation/StoreNavigation.kt
@@ -16,7 +16,7 @@ fun NavController.navigateToStore(
 
 fun NavGraphBuilder.storeGraph(
     navigateToUp: () -> Unit,
-    navigateToProfileEdit: () -> Unit,
+    navigateToProfileEdit: (Long) -> Unit,
     navigateToProductDetail: (Long) -> Unit,
     navigateToStoreReport: (Long) -> Unit,
     modifier: Modifier = Modifier,
